+++ pkg/kubelet/src/lib.rs
@@ -24,7 +24,6 @@ pub mod pod_manager;
 pub mod runtime;
 pub mod server;
 pub mod vm_manager;
-pub mod vm_net;
 pub mod vm_migrate;
 pub mod vm_runtime;